Changed string from of causal estimator when p=0.0 so instead it prints p< (1/number of simulations). Might consider writing an attribute of CausalEstimator to output the p value so this is done in all circumstances.
Changed a line in test_significance that threw an exception when the estimated value was lower than the median estimate of the simulations.

Permit passing of number of simulations using the method_params parameter (useful because the default 1000 simulations for propensity score matching can take several minutes and because test_significance is not usually called manually).