remove basic logging configuration
When `basicConfig` is called, logging is set up automatically and unless
a `logging.NullHandler` instance is included with each call it becomes
very difficult to control logging from DoWhy. This change removes all
calls to `basicConfig` and adds a `NullHandler` to the root DoWhy
logger. Now, users of DoWhy can configure their logging preference
through something like `logging.config.dictConfig` with a reference to
`dowhy` in the `loggers` section of the configuration.

Closes microsoft/dowhy#230