Revise and refactor kernel tests
Simplified kernel tests. Now, the inputs are converted to numeric vectors and one specific kernel is used for all variables (instead of a different kernel per feature). This simplifies the customization of the utilized kernel and using the lambda function still offers the flexibility to concatenate different kernels into one function.