Revise and simplify approx_kernel_based
Now, the inputs are converted to numeric vectors and one specific approximation is used is used for all variables (instead of a different approximation per feature). This simplifies the customization of the utilized kernel approximation method and using the lambda function still offers the flexibility to concatenate different approximations into one function.