Change one-hot-encoding behaviour in gcm module
Before, one dimension is dropped in the encoding. However, this requires to assume that there are no unknown categories, since these would be mapped to a zero vector as well, which then coincides with one of the categories. Now, there are as many dimensions as categories, which allow to map unknown categories to a zero vector.

Signed-off-by: Patrick Bloebaum <bloebp@amazon.com>