Add catboost encoder for modeling FCMs in gcm module
Before, one hot encoding was used to convert categorical features to a numeric vector. However, this doesn't scale well when the number of categories is high, i.e., we get very high dimensional vectors. Instead, we now use a target encoder based on the Catboost encoding when the number of categories is larger than a certain threshold.

Signed-off-by: Patrick Bloebaum <bloebp@amazon.com>