Explose individual interventional outcomes from RegressionEstimator _do operator.
Original implementation of _do for RegressionEstimators calculated individual outcomes for all dataframe rows, and then returned the mean(). However, for many use-cases it is helpful to have the individual outcomes for further analysis. This commit moves the existing implementation (without changes) to a new function interventional_outcomes() which returns all outcomes. The implementation of _do() now only calls mean() on the returned values. The behaviour of _do() is unchanged.

Signed-off-by: drawlinson <dave@agi.io>